First pass at data structure editor
Add an "edit" button to collection views. Users can edit and save EDN
and the underlying datastructures will be updated.

I had to be careful with the save process because

(om/update! data new-data)

triggers a re-mount of the component, which caused an error when we
tried to set state on the newly unmounted component.